fix(governance): the edit drawer follows its page through the Inventory rename
Main renamed `ingestion-sources.tsx` to `inventory.tsx` when the nav unified
Tool Tiles and Sources (15d54d1005). Rebasing onto that moved this branch's
edits into the renamed file correctly — git follows content — but the import
specifiers naming the old path are strings, and nothing rewrites a string.

So the detail page and all four edit-path test files still asked for
`./ingestion-sources`, a module no longer there. Typecheck reported it as five
unresolvable imports plus one implicit `any`, the last of which was only a
knock-on: with the module unresolved there was no prop type left to infer
`onSubmit`'s parameter from.

The kind of break a rebase cannot flag, because at no point were two versions
of a line in conflict.
