+++ platform/app/ee/governance/dashboard/pages/__tests__/editPullSourceConfig.builders.unit.test.ts
@@ -26,7 +26,7 @@ import {
   buildAnthropicAdminPullConfig,
   buildEditedParserConfig,
   buildEditSubmission,
-} from "../ingestion-sources";
+} from "../inventory";
 import { composer } from "./editPullSourceConfig.fixture";
 
 // `resolvePullConfig` toasts the offending field when a pull config will not
